Removed caching mechanism.
The design is badly though and currently not used by customers.
The current design doesn't fully handle management of the cache
and is hence half managed with responsibility of the users. A new
design will be developed with a fully managed design (in the meantime
it would be simpler for users to just manage their caching independently).