Use keyword nullptr where possible
Replace every usage of 0x0 or NULL with nullptr, which is available
since C++11.  NULL, or 0x0, or nullptr, or even 0 was used over the
entiry code base before for expressing the concept of a null pointer.
Use the C++ special word nullptr instead to unify the code base and
minimize friction with expectations of modern C++ developers.

References: #241
Signed-off-by: Alexander Dahl <post@lespocky.de>