file_response: Change default argument for content-type
The default value for content-type should be as broad and unspecific as
possible, to allow using file_response and don't care about the
content-type. Mozilla recommends to use application/octet-stream for
data with unknown type.

text/plain is too specific for using files of unknown type with the
file_response class, and could lead to browsers misinterpreting, showing
garbage, or even breaking sites.

Link: https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types
Fixes: #248
Signed-off-by: Alexander Dahl <post@lespocky.de>