file_response: Test on regular file
It was possible before to pass a path to a directory, or a to a device
file, or to basically any path.  `open()` would happily open it.
In case of a directory, `lseek()` returns LONG_MAX (0x7FFFFFFFFFFFFFFF)
and `MHD_create_response_from_fd()` is invoked.  To avoid such nonsense,
we test the path now and allow regular files only.

References: #255