Fix the webserver class to return 500 on invalid MHD Response
This particularly applies in case of a file_response (in case the file doesn't exist or is a directory).

* webserver: Return 500 on invalid MHD response

If `http_response::get_raw_response()` returns nullptr instead of a
valid `MHD_Response*` for whatever reason, that pointer would be passed
on to `http_response::decorate_response()` and
`http_response::enqueue_response()` eventually, leading to different API
calls to libmicrohttpd with NULL as argument `struct MHD_Response
*response`.  MHD does not guarantee any form of behaviour for invalid
input, so we have to consider it undefined behaviour and avoid passing
invalid input to MHD.

HTTP status 500 (Internal Server Error) is returned for consistency with
surrounding error handling, we don't know what caused
`http_response::get_raw_response()` to return nullptr, so we can not
give a better answer here.

Fixes: #255

* file_response: Add return value checks

Both `open()` and `lseek()` might fail depending on how `filename` is
set in object of class `httpserver::file_response()`.  In the case of a
missing file *fd* got -1 and lseek set *size* (which had the wrong type
btw.) to 0xffffffff aka (off_t) -1.  Passing an invalid file descriptor
and a massively huge size value on to `MHD_create_response_from_fd()`
might lead to unpredictable results depending how well libmicrohttpd
treats such invalid values.

Note: Before f9b769106435 ("Use MHD_create_response_from_fd for files")
`httpserver::http::load_file()` was used, which throws an exception in
case a file can not be opened successfully.  That exception would have
lead to returning HTTP status 500 (Internal Server Error).

References: #255

* test: Add unit test for missing file response

The constructor of class `httpserver::file_response` can be called with
a `filename` pointing into the void, to a file which does not actually
exist.  The webserver should fail predictably in that case.

References: #255

* file_response: Test on regular file

It was possible before to pass a path to a directory, or a to a device
file, or to basically any path.  `open()` would happily open it.
In case of a directory, `lseek()` returns LONG_MAX (0x7FFFFFFFFFFFFFFF)
and `MHD_create_response_from_fd()` is invoked.  To avoid such nonsense,
we test the path now and allow regular files only.

References: #255

* file_response: Add API doc to constructor

This documents a possible pitfall for users when passing filename of not
existing files.

References: #255

* test: Add unit test for file_response pointing to directory

The constructor of class `httpserver::file_response` can be called with
a `filename` pointing to a directory instead of a regular file.
The webserver should fail predictably in that case.

References: #255

* readme: Document requirements and behaviour of file_response

Suggested-by: Sebastiano Merlino <sebastiano@hey.com>
References: #255