Made http_utils::generate_random_upload_filename plattform independant
Mkstemp is only available on windows. As there is no suitable plattform independant function
to generate a unique filename within a specified directory, use mkstemp on linux and a
combination of _mktemp_s and _sopen_s on windows.
Furthermore the path separator differs on linux and windows so use a static const char
which provides the appropriate path separator

Additionally within this patch the function http_utils::generate_random_upload_filename
uses an exception to reports errors rather than a empty string as return code