Added an example for file upload
file_upload.cpp is a very basic example for a file upload server.
It features a simple HTML input form with file upload (which supports multiple files).
The uploaded files will be stored to the directory, which has to be given as
command line parameter to the example program.
All uploaded files will be printed in a table including their original name,
the generated random filepath and the size.