Changed file map to be a map of keys with a nested map of files
There is the possibility to upload the same file with two different keys
or to upload multiple (different) files with the same key within the
same request.
The current solution can only handle the second use case but not the first.
This patch changes the file map to use a nested map of files within
a map of keys to support both use cases.