Removed strlen from generate_random_upload_filename
As Codacy static code analysis warns about the safety issue of strlen
if a given string is not terminated, this is removed.
The solution uses the size of the original string, which is just duplicated
to a char* as the used function _mktemp_s requires a char* which will
be changed by the function and the changed result is needed afterwards.