Setters of class file_info are private and class webserver is a friend of this class
As the file_info has to be changed (especially the file_size) in multiple iterations
of the post iterator, this class cannot be fully immutable.
To realise a interface without setters to the outside of the library these
setters are private now. The only other classe, which uses these setters
is the class webserver and is therefor declared as friend to the file_info class.