Don't create zero length files if no file is uploaded
If an upload field is submitted without a file selected, the previous
implementation would still create a random file with zero length.
This is due to the fact, that the key would be present in the content,
but the filename is empty.

So a simple check, whether the filename is really present is introduced.

This is done by simply looking at the very first sign of the const char*
(which is given from libmicrohttpd). As strlen is considered unsafe and
there is already a check, whether the const char* is not a nullptr, it is
always safe to dereference and inspect the first char.