Close open upload file after post processor is done
If the file is not closed after the post processor is done, the
last info of the file might not be written to disk until the calling
application accesses the file. This could easily be reproduceable with
uploading very small files ( < 100 Bytes). In this case the content
of the file was missing for the calling process.