test: integ: file_upload: Stop webserver earlier
The webserver runs in a different thread than the test, therefor
deleting the uploaded files and checking if they are deleted might lead
into a race condition, and thus into tests failing sometimes, but not
always.

Moving the webserver stop and destructor behind the curl action should
be safe, every information needed for the test is copied to or in the
resource handler.