Cleanup files uploaded on request deletion (#266)
* http_request: Remove uploaded files in destructor

If the user needs those files, she should copy/move theme in the request
handler.

References: #264
Signed-off-by: Alexander Dahl <ada@thorsis.com>

* test: integ: file_upload: Remove redundant file delete

Not needed anymore. http_request destructor does cleanup now.

* test: integ: file_upload: Check uploaded files are deleted

* test: integ: file_upload: Stop webserver earlier

The webserver runs in a different thread than the test, therefor
deleting the uploaded files and checking if they are deleted might lead
into a race condition, and thus into tests failing sometimes, but not
always.

Moving the webserver stop and destructor behind the curl action should
be safe, every information needed for the test is copied to or in the
resource handler.