fix #292 handle argument keys with multiple values
This change allows the server to handle argument keys with multiple values
e.g. url?arg=param&arg=what

Previous, the arg key would just be overwritten with the next value. This
change adds a http_arg_value type to house all the values for a given key.

Adjust existing get_arg(s) methods, and add some get_arg_flat type
methods that still just return the first value for a given key.

Add a test case for multiple valued keys.

Add a test case for a large file upload that triggers chunking behavior
of MHD. This causes the server to concatenate file chunks within the
first value of the given arg key, so the test ensures that this
behavior is not affected by the changes.