Update webserver.cpp
Address an error that _key_ in the static PSK handler function is assigned a value that is never used in the case if the Gnu TLS libary is not present, i.e. HAVE_GNUTLS has not been defined.

The intended behavior is such that calling the static callback handler function of the webserver guarantees that the user registered callback function is invoked.